refact : ACE 판단 책임 CardNumber로 이동
