+++ src/main/kotlin/blackjack/domain/card/CardNumber.kt
@@ -1,5 +1,8 @@
 package blackjack.domain.card
 
+import blackjack.domain.CardNumberCalculator
+import blackjack.domain.RuleChecker
+
 enum class CardNumber(val value: Int) {
     A(1),
     TWO(2),
@@ -23,5 +26,17 @@ enum class CardNumber(val value: Int) {
             A -> "A"
             else -> value.toString()
         }
+    }
+
+    companion object {
+        fun proceedAceNumber(sum: Int): Int {
+            if (sum + ACE_MAXINUM <= RuleChecker.CONDITION_TO_WIN_BLACK_JACK) {
+                return ACE_MAXINUM
+            }
+            return ACE_MINIMUM
+        }
+
+        private const val ACE_MAXINUM = 11
+        private const val ACE_MINIMUM = 1
     }
 }
